requirements.txt
streamlit
torch
torchvision
pandas
numpy
Pillow
openpyxl
scikit-learn
opencv-python
